modules segregation,improved auth,lazyloading,ui Improved

+++ src/app/recruiter/recruiter.component.ts
@@ -1,12 +1,13 @@
 import {  Component,OnInit, ViewChild } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-
 import { AddJobsComponent } from './add-jobs/add-jobs.component';
-// import { ServerService } from '../server.service';
 import { MatTableDataSource } from '@angular/material/table';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatPaginator } from '@angular/material/paginator';
 import { ServerService } from 'src/services/server.service';
+import { AlertComponent } from './alert/alert.component';
+import { ShowoneComponent } from '../shared/showone/showone.component';
+import { ViewSingleJob } from './recruiter.model'  // interface
 
 
 @Component({
@@ -34,59 +35,70 @@ export class RecruiterComponent implements OnInit {
   }
   email = this.getEmail()
 
-  response: any[] = [];
+  response: ViewSingleJob[] = [];
 
+  // this will the call the api will get the jobs created by specific recruiter
   loadData(){
     this.apiCalls.showJobs(this.email).subscribe((res)=>{
       this.response = res
+      // console.log(this.response);
       this.dataSource = new MatTableDataSource(this.response);
       this.dataSource.paginator = this.paginator;
     
     })
   }
+
   ngOnInit(): void {
     this.loadData()
   }
 
+
+  // table columns definitions
   displayedColumns: string[] = ['id', 'jobtitle', 'salary', 'seekers', 'actions'];
 
- 
-  deleteJob(id:any){
-    this.apiCalls.deleteJob(id).subscribe((res)=>{
-      this.response = res
-      this.loadData()
-    })
 
+  // single job view at a time and seekers list
+  onView(id:number){
+    let ViewSingleJob !: ViewSingleJob
+    this.apiCalls.viewJob(id).subscribe((res)=>{
+      ViewSingleJob = res
+      // this.dialog.open(ViewjobComponent,{data:ViewSingleJob})
+      this.dialog.open(ShowoneComponent,{data:ViewSingleJob})
+    })
   }
 
   
-  viewReponse:any
-  exists:boolean = false
-  onView(id:number){
-    this.exists = true
-    // console.log(id)
-    this.apiCalls.viewJob(id).subscribe((res)=>{
-      this.viewReponse = res
-    })
-  }
 
-  viewClose(){
-    this.exists = false
-  }
 
-  openDialog(enterAnimationDuration: string, exitAnimationDuration: string): void {
-    const ref =   this.dialog.open(AddJobsComponent, {
-      width: '250px',
-      enterAnimationDuration,
-      exitAnimationDuration,
-    });
-
+  // add jobs dialog box
+  openDialogAddjob() {
+    const ref =   this.dialog.open(AddJobsComponent);
     ref.afterClosed().subscribe(()=>{
-      // console.log("from ref")
       this.loadData()
     })
- 
   }
+
+  // on clicking on deleting jobs
+  deleteJob(id:number,jobtitle:string){
+    this.OpenDialogAlert(id,jobtitle)
+  }
+
+  OpenDialogAlert(id: number,jobtitle: string){
+    const alertDialog = this.dialog.open(AlertComponent,
+      { data:{id,jobtitle},
+        width: '30vw',
+        height: '30vh',
+      })
+    alertDialog.afterClosed().subscribe((result)=>{
+      if(result === 'yes' ){
+        this.apiCalls.deleteJob(id).subscribe((res)=>{
+            this.response = res
+            this.loadData()
+          })
+      }      
+    })
+  }
+
 
 }
